requirements.txt
openai-whisper
flask
flasgger
flask-cors